chore: update tui-sandbox from 2.3.0 to 4.0.0
https://github.com/mikavilpas/tui-sandbox/pull/119
