+++ packages/integration-tests/MyTestDirectory.ts
@@ -11,7 +11,7 @@
 import { z } from "zod"
 
 export const MyTestDirectorySchema = z.object({
-  name: z.literal("test-environment"),
+  name: z.literal("test-environment/"),
   type: z.literal("directory"),
   contents: z.object({
     "file.txt": z.object({
@@ -33,7 +33,7 @@ export const MyTestDirectorySchema = z.object({
       stem: z.literal("multicursor-file."),
     }),
     "other-subdirectory": z.object({
-      name: z.literal("other-subdirectory"),
+      name: z.literal("other-subdirectory/"),
       type: z.literal("directory"),
       contents: z.object({
         "other-sub-file.txt": z.object({
@@ -45,11 +45,11 @@ export const MyTestDirectorySchema = z.object({
       }),
     }),
     routes: z.object({
-      name: z.literal("routes"),
+      name: z.literal("routes/"),
       type: z.literal("directory"),
       contents: z.object({
         "posts.$postId": z.object({
-          name: z.literal("posts.$postId"),
+          name: z.literal("posts.$postId/"),
           type: z.literal("directory"),
           contents: z.object({
             "adjacent-file.txt": z.object({
@@ -75,7 +75,7 @@ export const MyTestDirectorySchema = z.object({
       }),
     }),
     subdirectory: z.object({
-      name: z.literal("subdirectory"),
+      name: z.literal("subdirectory/"),
       type: z.literal("directory"),
       contents: z.object({
         "subdirectory-file.txt": z.object({
