+++ packages/integration-tests/cypress.config.ts
@@ -2,6 +2,8 @@ import { defineConfig } from "cypress"
 
 export default defineConfig({
   e2e: {
+    baseUrl: "http://localhost:3000",
+    experimentalRunAllSpecs: true,
     retries: {
       runMode: 2,
       openMode: 0,
