+++ packages/integration-tests/cypress/e2e/my-live-grep.cy.ts
@@ -1,6 +1,6 @@
 describe("my_live_grep", () => {
   it("can grep in the project (start in normal mode)", () => {
-    cy.visit("http://localhost:5173")
+    cy.visit("/")
     cy.startNeovim().then((dir) => {
       // wait until text on the start screen is visible
       cy.contains("If you see this text, Neovim is ready!")
@@ -19,7 +19,7 @@ describe("my_live_grep", () => {
   })
 
   it("can grep in the project (start in insert mode)", () => {
-    cy.visit("http://localhost:5173")
+    cy.visit("/")
     cy.startNeovim().then((dir) => {
       // wait until text on the start screen is visible
       cy.contains("If you see this text, Neovim is ready!")
