+++ packages/integration-tests/cypress/support/commands.ts
@@ -1,15 +1,35 @@
 /* eslint-disable @typescript-eslint/no-namespace */
 /// <reference types="cypress" />
 
-import "../../client/__global.ts"
-import type { NeovimContext } from "../../client/__global.ts"
-import type { MyStartNeovimServerArguments } from "../../client/neovim-client.ts"
+import type {
+  MyTestDirectory,
+  MyTestDirectoryFile,
+} from "../../MyTestDirectory"
+
+export type NeovimContext = {
+  contents: MyTestDirectory
+  rootPathAbsolute: string
+}
+
+declare global {
+  interface Window {
+    startNeovim(
+      startArguments?: MyStartNeovimServerArguments,
+    ): Promise<NeovimContext>
+  }
+}
+
+type MyStartNeovimServerArguments = {
+  filename?:
+    | MyTestDirectoryFile
+    | { openInVerticalSplits: MyTestDirectoryFile[] }
+}
 
 Cypress.Commands.add(
   "startNeovim",
   (startArguments?: MyStartNeovimServerArguments) => {
-    cy.window().then((win) => {
-      return win.startNeovim(startArguments)
+    cy.window().then(async (win) => {
+      return await win.startNeovim(startArguments)
     })
   },
 )
