--- packages/integration-tests/index.html
@@ -1,11 +1,0 @@
-<!doctype html>
-<html lang="en">
-  <head>
-    <meta charset="UTF-8" />
-    <title>integration tests</title>
-  </head>
-  <body>
-    <div id="app"></div>
-    <script type="module" src="./client/neovim-client.ts"></script>
-  </body>
-</html>
